backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
toml
websockets
huggingface_hub
onnxruntime
Pillow
numpy
pandas
diffusers
accelerate
transformers
ftfy
albumentations
einops
colorama
tqdm
lion-pytorch
lycoris-lora
